refactor(self-serve): split single massive Products class into multiple, for readability (#233)
* refactor(self-serve): split single massive Products class into multiple, for readability

* refactor(self-serve): consolidate is_active method and init subclasses conditionally

* refactor: consolidate active checks for both products and featured classes

* refactor: only run featured methods if environment variable is true

* fix: missing dependency in UI subclass

* refactor: cleaner class and file naming conventions using namespace

* fix: update class name
