+++ includes/class-api.php
@@ -9,11 +9,11 @@
 
 namespace Newspack_Listings;
 
-use \Newspack_Listings\Newspack_Listings_Core as Core;
-use \Newspack_Listings\Newspack_Listings_Blocks as Blocks;
-use \Newspack_Listings\Newspack_Listings_Featured as Featured;
-use \Newspack_Listings\Newspack_Listings_Taxonomies as Taxonomies;
-use \Newspack_Listings\Utils as Utils;
+use \Newspack_Listings\Core;
+use \Newspack_Listings\Blocks;
+use \Newspack_Listings\Featured;
+use \Newspack_Listings\Taxonomies;
+use \Newspack_Listings\Utils;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -21,26 +21,26 @@ defined( 'ABSPATH' ) || exit;
  * API class.
  * Sets up API endpoints and handlers for listings.
  */
-final class Newspack_Listings_Api {
+final class Api {
 	/**
 	 * REST route namespace.
 	 *
-	 * @var Newspack_Listings_Api
+	 * @var Api
 	 */
 	protected static $namespace = 'newspack-listings/v1';
 
 	/**
 	 * The single instance of the class.
 	 *
-	 * @var Newspack_Listings_Api
+	 * @var Api
 	 */
 	protected static $instance = null;
 
 	/**
-	 * Main Newspack_Listings_Api instance.
-	 * Ensures only one instance of Newspack_Listings_Api is loaded or can be loaded.
+	 * Main Api instance.
+	 * Ensures only one instance of Api is loaded or can be loaded.
 	 *
-	 * @return Newspack_Listings_Api - Main instance.
+	 * @return Api - Main instance.
 	 */
 	public static function instance() {
 		if ( is_null( self::$instance ) ) {
@@ -197,7 +197,7 @@ final class Newspack_Listings_Api {
 		);
 
 		// Get and set listing priority level.
-		if ( defined( 'NEWSPACK_LISTINGS_SELF_SERVE_ENABLED' ) && NEWSPACK_LISTINGS_SELF_SERVE_ENABLED ) {
+		if ( Featured::is_active() ) {
 			register_rest_route(
 				self::$namespace,
 				'priority',
@@ -205,7 +205,7 @@ final class Newspack_Listings_Api {
 					[
 						'methods'             => \WP_REST_Server::READABLE,
 						'callback'            => [ __CLASS__, 'get_priority' ],
-						'permission_callback' => '__return_true',
+						'permission_callback' => [ __CLASS__, 'api_permissions_check' ],
 					],
 				]
 			);
@@ -664,4 +664,4 @@ final class Newspack_Listings_Api {
 	}
 }
 
-Newspack_Listings_Api::instance();
+Api::instance();
