+++ includes/class-block-patterns.php
@@ -9,8 +9,8 @@
 
 namespace Newspack_Listings;
 
-use \Newspack_Listings\Newspack_Listings_Core as Core;
-use \Newspack_Listings\Utils as Utils;
+use \Newspack_Listings\Core;
+use \Newspack_Listings\Utils;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -18,7 +18,7 @@ defined( 'ABSPATH' ) || exit;
  * Blocks class.
  * Sets up custom blocks for listings.
  */
-final class Newspack_Listings_Block_Patterns {
+final class Block_Patterns {
 	/**
 	 * Slug for the block pattern category.
 	 */
@@ -232,4 +232,4 @@ final class Newspack_Listings_Block_Patterns {
 	}
 }
 
-Newspack_Listings_Block_Patterns::instance();
+Block_Patterns::instance();
