+++ includes/utils.php
@@ -7,8 +7,8 @@
 
 namespace Newspack_Listings\Utils;
 
-use \Newspack_Listings\Newspack_Listings_Core as Core;
-use \Newspack_Listings\Newspack_Listings_Featured as Featured;
+use \Newspack_Listings\Core;
+use \Newspack_Listings\Featured;
 
 /**
  * Sanitize an array of text or number values.
@@ -411,7 +411,7 @@ function get_term_classes( $post_id = null ) {
 	);
 
 	// If the item is featured, append class names for its featured status and priority level.
-	if ( defined( 'NEWSPACK_LISTINGS_SELF_SERVE_ENABLED' ) && NEWSPACK_LISTINGS_SELF_SERVE_ENABLED ) {
+	if ( Featured::is_active() ) {
 		$base_classes = Featured::add_featured_classes( $base_classes );
 	}
