+++ src/blocks/list-container/view.php
@@ -8,8 +8,8 @@
 
 namespace Newspack_Listings\Curated_List_Container_Block;
 
-use \Newspack_Listings\Newspack_Listings_Core as Core;
-use \Newspack_Listings\Utils as Utils;
+use \Newspack_Listings\Core;
+use \Newspack_Listings\Utils;
 
 /**
  * Dynamic block registration.
