+++ src/templates/self-serve-form.php
@@ -6,9 +6,9 @@
  * @package WordPress
  */
 
-use \Newspack_Listings\Newspack_Listings_Core as Core;
-use \Newspack_Listings\Newspack_Listings_Settings as Settings;
-use \Newspack_Listings\Utils as Utils;
+use \Newspack_Listings\Core;
+use \Newspack_Listings\Settings as Settings;
+use \Newspack_Listings\Utils;
 
 call_user_func(
 	function( $data ) {
